Refactoring and improving menu usage
