+++ src/components/navigation/MainMenu.types.ts
@@ -23,4 +23,5 @@ export type SubMenuItem = {
   title: string
   href?: MenuItemHrefType
   customOnClickBehavior?: () => void
+  subItems?: Array<SubMenuItem>
 }
